couche metier in progress

+++ src/main/java/myboot/app1/web/PersonRestController.java
@@ -5,39 +5,62 @@ import myboot.app1.dao.PersonRepository;
 import myboot.app1.model.Person;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/persons")
 public class PersonRestController {
 
     @Autowired
     PersonRepository PR;
 
 
-    @GetMapping("/persons")
-    public Iterable<Person>  getPersons() {
+
+    @GetMapping("/showAll")
+    public Iterable<Person> getPersons() {
         return PR.findAll();
     }
 
-    @GetMapping("/persons/{id}")
+    @GetMapping("/showPerson/{id}")
     public Person getPerson(@PathVariable int id) {
         return PR.findById(id).get();
     }
 
-    @DeleteMapping("/persons/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    void deletePerson(@PathVariable int id) {
-        PR.deleteById(id);
+    @DeleteMapping("/deletePerson/{id}")
+    public ResponseEntity<Boolean> deletePerson(@PathVariable (value="id") int id){
+        if(PR.existsById(id)){
+            PR.deleteById(id);
+            return new ResponseEntity<>(true,HttpStatus.OK);
+
+        }
+        else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
     }
 
-    @PostMapping("/persons")
+    @PostMapping("/newPerson")
     public Person postPerson(@RequestBody @Valid Person p) {
         PR.save(p);
         return p;
     }
+
+    @PutMapping("/updatePerson")
+    public ResponseEntity<Person> updatePerson (@RequestBody Person p){
+
+        if(p.getFirstName()==null) throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"Firstname missed !");
+        if(p.getLastName()==null)  throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"Firstname missed !");
+        else {
+            return new ResponseEntity(PR.save(p), HttpStatus.OK);
+        }
+
+    }
+
+
